+++ decidim-surveys/spec/lib/decidim/surveys/component_spec.rb
@@ -6,6 +6,7 @@ describe "Surveys component" do # rubocop:disable RSpec/DescribeClass
   subject { component }
 
   let(:component) { create :surveys_component }
+  let(:new_component) { create :surveys_component }
 
   describe "before_destroy hooks" do
     context "when there are no answers" do
@@ -26,6 +27,16 @@ describe "Surveys component" do # rubocop:disable RSpec/DescribeClass
           "Can't destroy this component when there are survey answers"
         )
       end
+    end
+  end
+
+  context "when copying component" do
+    it "does not raise any error" do
+      expect { subject.manifest.run_hooks(:copy, old_component: component, new_component: new_component) }.not_to raise_error
+    end
+
+    it "create a survey component" do
+      expect { subject.manifest.run_hooks(:copy, old_component: component, new_component: new_component) }.to change { Decidim::Surveys::Survey.count }.by(1)
     end
   end
 end